배포 작업 진행 및 백엔드, 프론트엔드 쪽 리팩토링 진행

+++ be/routes/contest.js
@@ -27,6 +27,10 @@ router.post('/add', async (req, res) => {
 });
 
 router.post('/delete', async (req, res) => {
+  if (!req.session.user) {
+    res.status(201).json(false);
+    return;
+  }
   const { contest_id } = req.body;
 
   const response = await contestModel.deleteOne({
